Move command builders in a different namespace
