+++ src/Core/Form/IdentifiableObject/CommandBuilder/Product/ProductCommandsBuilder.php
@@ -26,42 +26,44 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Command\Builder;
+namespace PrestaShop\PrestaShop\Core\Form\IdentifiableObject\CommandBuilder\Product;
 
-use PrestaShop\PrestaShop\Core\Domain\Product\Command\UpdateProductBasicInformationCommand;
-use PrestaShop\PrestaShop\Core\Domain\Product\QueryResult\ProductType;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 
 /**
- * Builder used to build UpdateProductBasicInformationCommand
+ * This class builds a collection of product commands based on the form data and a list of ProductCommandBuilderInterface
  */
-class BasicInformationCommandBuilder implements ProductCommandBuilderInterface
+class ProductCommandsBuilder
 {
     /**
-     * {@inheritdoc}
+     * @var ProductCommandBuilderInterface[]
      */
-    public function buildCommand(ProductId $productId, array $formData)
+    private $commandBuilders;
+
+    /**
+     * @param ProductCommandBuilderInterface[] $commandBuilders
+     */
+    public function __construct(array $commandBuilders)
     {
-        if (!isset($formData['basic'])) {
-            return null;
+        $this->commandBuilders = $commandBuilders;
+    }
+
+    /**
+     * @param ProductId $productId
+     * @param array $formData
+     *
+     * @return ProductCommandCollection
+     */
+    public function buildCommands(ProductId $productId, array $formData): ProductCommandCollection
+    {
+        $commands = new ProductCommandCollection();
+        foreach ($this->commandBuilders as $commandBuilder) {
+            $command = $commandBuilder->buildCommand($productId, $formData);
+            if (null !== $command) {
+                $commands->add($command);
+            }
         }
 
-        $basicData = $formData['basic'];
-        $command = new UpdateProductBasicInformationCommand($productId->getValue());
-
-        if (isset($basicData['name'])) {
-            $command->setLocalizedNames($basicData['name']);
-        }
-        if (isset($basicData['type'])) {
-            $command->setVirtual($basicData['type'] === ProductType::TYPE_VIRTUAL);
-        }
-        if (isset($basicData['description'])) {
-            $command->setLocalizedDescriptions($basicData['description']);
-        }
-        if (isset($basicData['description_short'])) {
-            $command->setLocalizedShortDescriptions($basicData['description_short']);
-        }
-
-        return $command;
+        return $commands;
     }
 }
