+++ src/PrestaShopBundle/Resources/config/services/core/form/form_data_handler.yml
@@ -182,4 +182,4 @@ services:
       class: 'PrestaShop\PrestaShop\Core\Form\IdentifiableObject\DataHandler\ProductFormDataHandler'
       arguments:
           - '@prestashop.core.command_bus'
-          - '@prestashop.adapter.product.command.builder'
+          - '@prestashop.core.form.command_builder.product_commands_builder'
