+++ tests/Unit/Core/Form/IdentifiableObject/CommandBuilder/Product/AbstractProductCommandBuilderTest.php
@@ -24,21 +24,40 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\Command\Builder;
+declare(strict_types=1);
 
+namespace Tests\Unit\Core\Form\IdentifiableObject\CommandBuilder\Product;
+
+use PHPUnit\Framework\TestCase;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
+use SebastianBergmann\Comparator\Factory as ComparatorFactory;
 
 /**
- * This interface is used by ProductCommandsBuilder each object which implements must build
- * a product command based on the input form data.
+ * Base class to test a product command builder
  */
-interface ProductCommandBuilderInterface
+abstract class AbstractProductCommandBuilderTest extends TestCase
 {
     /**
-     * @param ProductId $productId
-     * @param array $formData
-     *
-     * @return mixed|null Returns null if the required data for the command is absent
+     * @var ProductId
      */
-    public function buildCommand(ProductId $productId, array $formData);
+    private $productId;
+
+    protected function setUp()
+    {
+        parent::setUp();
+        $factory = ComparatorFactory::getInstance();
+        $factory->register(new ProductCommandComparator());
+    }
+
+    /**
+     * @return ProductId
+     */
+    protected function getProductId(): ProductId
+    {
+        if (null === $this->productId) {
+            $this->productId = new ProductId(42);
+        }
+
+        return $this->productId;
+    }
 }
