+++ tests/Unit/Core/Form/IdentifiableObject/CommandBuilder/Product/ProductCommandsBuilderTest.php
@@ -26,12 +26,12 @@
 
 declare(strict_types=1);
 
-namespace Tests\Unit\Core\Domain\Product\Builder;
+namespace Tests\Unit\Core\Form\IdentifiableObject\CommandBuilder\Product;
 
-use PrestaShop\PrestaShop\Core\Domain\Product\Command\Builder\ProductCommandBuilderInterface;
-use PrestaShop\PrestaShop\Core\Domain\Product\Command\Builder\ProductCommandCollection;
-use PrestaShop\PrestaShop\Core\Domain\Product\Command\Builder\ProductCommandsBuilder;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
+use PrestaShop\PrestaShop\Core\Form\IdentifiableObject\CommandBuilder\Product\ProductCommandBuilderInterface;
+use PrestaShop\PrestaShop\Core\Form\IdentifiableObject\CommandBuilder\Product\ProductCommandCollection;
+use PrestaShop\PrestaShop\Core\Form\IdentifiableObject\CommandBuilder\Product\ProductCommandsBuilder;
 
 class ProductCommandsBuilderTest extends AbstractProductCommandBuilderTest
 {
